+++ lib/nerves_grove/i2c.ex
@@ -1,5 +1,0 @@
-# This is free and unencumbered software released into the public domain.
-
-defmodule Nerves.Grove.I2C do
-  # TODO
-end